Moved config files to detecated folder
