+++ src/main/java/com/heroku/backend/config/ValueConfig.java
@@ -1,10 +1,10 @@
-package com.heroku.backend;
+package com.heroku.backend.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class ValueConfiguration {
+public class ValueConfig {
 
     public static String mongoDbUri;
     public static String authenticationToken;
@@ -13,22 +13,22 @@ public class ValueConfiguration {
 
     @Value("${mongodb.uri}")
     public void setMongoDbUri(String mongoDbUri) {
-        ValueConfiguration.mongoDbUri = mongoDbUri;
+        ValueConfig.mongoDbUri = mongoDbUri;
     }
 
     @Value("${authentication.user}")
     public void setAuthenticationToken(String authenticationToken) {
-        ValueConfiguration.authenticationToken = authenticationToken;
+        ValueConfig.authenticationToken = authenticationToken;
     }
 
     @Value("${users.jan.base64}")
     public void setBase64FirstUser(String base64FirstUser) {
-        ValueConfiguration.base64FirstUser = base64FirstUser;
+        ValueConfig.base64FirstUser = base64FirstUser;
     }
 
     @Value("${users.lea.base64}")
     public void setBase64SecondUser(String base64SecondUser) {
-        ValueConfiguration.base64SecondUser = base64SecondUser;
+        ValueConfig.base64SecondUser = base64SecondUser;
     }
 
 }
